// frontend/src/lib/api/generated/model/departmentServiceDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ระบบรายงานค่าสาธารณูปโภค กรมประมง API
 * OpenAPI spec version: 2.0.0
 */

export interface DepartmentServiceDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  createdAt: string;
  departmentId: string;
  departmentName?: string;
  id: string;
  locationType?: string;
  phoneOwnerName?: string;
  phoneOwnerPosition?: string;
  phoneReimbursementLimit?: number;
  provider: string;
  serviceNumber: string;
  utilityType: string;
}